requirements.txt
Django==4.2.23
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
drf-spectacular==0.27.1
python-decouple==3.8
whitenoise==6.6.0
dj-database-url==2.1.0
boto3==1.34.0
botocore==1.34.0
Pillow==10.1.0
django-storages==1.14.2
requests==2.31.0
setuptools>=65.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
